Move LocalFeedLoader and FeedStore collaborator to their own files in production
